requirements.txt
aiofiles==24.1.0
aiogram==3.13.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.8
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.7.0
asyncpg==0.29.0
attrs==24.2.0
beautifulsoup4==4.12.3
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
feedparser==6.0.11
frozenlist==1.4.1
greenlet==3.1.1
idna==3.10
kombu==5.4.2
magic-filter==1.0.12
multidict==6.1.0
prompt_toolkit==3.0.48
psycopg2==2.9.9
pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
redis==5.1.1
requests==2.32.3
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.6
SQLAlchemy==2.0.35
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
yarl==1.13.1
